+++ InfiniteRoleplay/Entities/Ponto.cs
@@ -7,9 +7,9 @@ namespace InfiniteRoleplay.Entities
     {
         public int Codigo { get; set; }
         public int Tipo { get; set; } = 0;
-        public double PosX { get; set; } = 0;
-        public double PosY { get; set; } = 0;
-        public double PosZ { get; set; } = 0;
+        public float PosX { get; set; } = 0;
+        public float PosY { get; set; } = 0;
+        public float PosZ { get; set; } = 0;
 
         [NotMapped]
         public TextLabel TextLabel { get; set; }
